requirements.txt
bunch>=1.0.1
boto>=2.31.1
PyYAML>=3.10
-e git+git@github.com:clearcare/cc_logger.git@0.0.4#egg=cc_logger
-e git+https://github.com/shokunin/python-logstash-formatter.git@3c3d218226#egg=logstash_formatter
